refactor: 💡 수정 유저가 보는 페이지 (메인, 매장, 메뉴) 에 대해 Controller 를 통합
